Move SchemaManager and SchemaBranch into separate files
