+++ backend/infrahub/core/merge.py
@@ -12,7 +12,7 @@ from infrahub.core.query.branch import (
 from infrahub.core.query.node import NodeDeleteQuery, NodeListGetInfoQuery
 from infrahub.core.registry import registry
 from infrahub.core.schema import GenericSchema, NodeSchema
-from infrahub.core.schema_manager import SchemaUpdateValidationResult
+from infrahub.core.schema.schema_branch import SchemaUpdateValidationResult
 from infrahub.core.timestamp import Timestamp
 from infrahub.core.utils import add_relationship, update_relationships_to
 from infrahub.exceptions import ValidationError
@@ -23,7 +23,8 @@ from .diff.branch_differ import BranchDiffer
 if TYPE_CHECKING:
     from infrahub.core.branch import Branch
     from infrahub.core.models import SchemaUpdateConstraintInfo, SchemaUpdateMigrationInfo
-    from infrahub.core.schema_manager import SchemaBranch, SchemaDiff
+    from infrahub.core.schema.manager import SchemaDiff
+    from infrahub.core.schema.schema_branch import SchemaBranch
     from infrahub.database import InfrahubDatabase
     from infrahub.services import InfrahubServices
