+++ backend/infrahub/core/protocols_base.py
@@ -7,7 +7,7 @@ from typing_extensions import Self
 if TYPE_CHECKING:
     from neo4j import AsyncResult, AsyncSession, AsyncTransaction, Record
 
-    from infrahub.core.schema_manager import SchemaBranch
+    from infrahub.core.schema.schema_branch import SchemaBranch
 
 # pylint: disable=redefined-builtin
