+++ backend/infrahub/core/validators/checker.py
@@ -8,7 +8,7 @@ from infrahub.message_bus.messages import MESSAGE_MAP, RESPONSE_MAP
 if TYPE_CHECKING:
     from infrahub.core.branch import Branch
     from infrahub.core.models import SchemaUpdateConstraintInfo
-    from infrahub.core.schema_manager import SchemaBranch
+    from infrahub.core.schema.schema_branch import SchemaBranch
     from infrahub.message_bus.messages.schema_validator_path import SchemaValidatorPathResponse
     from infrahub.services import InfrahubServices
