+++ backend/tests/helpers/test_app.py
@@ -17,7 +17,8 @@ from infrahub.core.initialization import (
     initialization,
 )
 from infrahub.core.schema import SchemaRoot, core_models, internal_schema
-from infrahub.core.schema_manager import SchemaBranch, SchemaManager
+from infrahub.core.schema.manager import SchemaManager
+from infrahub.core.schema.schema_branch import SchemaBranch
 from infrahub.core.utils import delete_all_nodes
 from infrahub.database import InfrahubDatabase
 from infrahub.server import app, app_initialization
