+++ backend/tests/unit/core/test_registry.py
@@ -1,7 +1,7 @@
 from infrahub.core.branch import Branch
 from infrahub.core.registry import registry
 from infrahub.core.schema import SchemaRoot, internal_schema
-from infrahub.core.schema_manager import SchemaManager
+from infrahub.core.schema.manager import SchemaManager
 from infrahub.database import InfrahubDatabase
