+++ pyproject.toml
@@ -267,8 +267,13 @@ module = "infrahub.core.schema.basenode_schema"
 ignore_errors = true
 
 [[tool.mypy.overrides]]
-module = "infrahub.core.schema_manager"
+module = "infrahub.core.schema.schema_branch"
 ignore_errors = true
+
+[[tool.mypy.overrides]]
+module = "infrahub.core.schema.manager"
+ignore_errors = true
+
 
 [[tool.mypy.overrides]]
 module = "infrahub.core.utils"
